+++ lib/native_admob_flutter.dart
@@ -15,3 +15,5 @@ export 'src/banner/controller.dart';
 export 'src/interstitial/interstitial.dart';
 
 export 'src/rewarded/rewarded.dart';
+
+export 'src/app_open/app_open.dart';